Add simple addFile function
Might seem a bit strange, but let's say you store all your JS files in a
directory. Now, you want to host your service-worker.js in the root (for
scope reasons). This will allow you to do addFile('/service-worker.js',
'./js/service-worker.js');

This basically is a little bit of url-rewriting for an individual file.

Patch is untested as on debian9 the string_view is in std::experimental.